refactor: refatoração do componente select-page
